@@ -4,33 +4,44 @@
   "private": true,
   "author": "Chau Le",
   "license": "ISC",
+  "type": "module",
   "dependencies": {
-    "@apollo/client": "^3.1.2",
-    "@prisma/client": "^2.4.1",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
-    "apollo-server": "^2.16.1",
+    "axios": "^1.4.0",
     "bcryptjs": "^2.4.3",
+    "body-parser": "^1.20.2",
     "bootstrap": "^4.5.2",
+    "connect-ensure-login": "^0.1.1",
     "cross-fetch": "^3.0.5",
+    "crypto": "^1.0.1",
+    "date-diff": "^1.0.2",
     "dotenv": "^8.2.0",
+    "ejs": "^3.1.9",
+    "express": "^4.18.2",
+    "express-mysql-session": "^3.0.0",
+    "express-session": "^1.17.3",
     "goodreads": "^0.1.0",
     "goodreads-api-node": "0.0.12",
-    "graphql": "^15.3.0",
     "jsonwebtoken": "^8.5.1",
+    "mysql": "^2.18.1",
+    "mysql2": "^3.5.1",
     "node-fetch": "^2.6.0",
-    "react": "^16.13.1",
+    "passport": "^0.6.0",
+    "passport-local": "^1.0.0",
+    "react": "^16.14.0",
     "react-dom": "^16.13.1",
     "react-infinite-scroller": "^1.2.4",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
+    "reactjs-popup": "^2.0.5",
     "xml2js": "^0.4.23"
   },
   "devDependencies": {
     "@prisma/cli": "^2.4.1",
     "dotenv": "^8.2.0",
-    "nodemon": "^2.0.4",
+    "nodemon": "^2.0.22",
     "xml-fetch": "^1.1.91"
   },
   "scripts": {
