@@ -1,15 +1,19 @@
-async function ownBooks(parent,args,context,info){
-    return await context.prisma.user.findOne({where:{id:parent.id}}).ownBooks()
-}
-async function sellTransactions(parent,args,context,info){
-    return await context.prisma.user.findOne({where:{id:parent.id}}).sellTransactions()
-}
-async function buyTransactions(parent,args,context,info){
-    return await context.prisma.user.findOne({where:{id:parent.id}}).buyTransactions()
+import {getQuery, getQueryOneOrNull} from '../database.js'
+
+export async function getUserByUsername(username) {
+    return getQueryOneOrNull(`SELECT * FROM users WHERE username = '${username}'`)
 }
 
-module.exports = {
-    ownBooks,
-    sellTransactions,
-    buyTransactions
+export async function getUserByEmail(email) {
+    return await getQueryOneOrNull(`SELECT * FROM users WHERE email = '${email}'`)
+}
+
+export async function getUsers() {
+    return getQuery(`SELECT * FROM users;`)
+}
+
+export async function createUser(user) {
+    const res =  getQuery(`INSERT INTO users (email, username, password) VALUES ('${user.email}', '${user.username}', '${user.password}');`)
+    console.log(res)
+    return res
 }
