@@ -1,150 +1,47 @@
 import React, { Fragment, useState, useEffect } from "react";
-import { gql, useQuery, useMutation } from "@apollo/client";
 import SellerBookBox from "../components/SellerBookBox"
 import { Redirect } from "react-router-dom";
 import FeedBox from "../components/FeedBox";
 import InfiniteScroll from "react-infinite-scroller";
 import SellPage from "./SellPage"
+import axios from "axios";
 
 
-const GET_POSTS = gql`
-  query feed($cursorId:Int,$take:Int){
-    feed(cursorId:$cursorId, take:$take){
-        cursorId
-    isNotEmpty
-    results{
-      id
-      price
-      dateForSale
-    	volumeIdGG
-      ownedBy{username}
-    }
-    
-  }
-    }   
-`;
-
-const BUY_BOOK = gql`
-    mutation buy($bookId:ID!){
-        buy(bookId:$bookId){
-            date
-            seller{
-            username
-            }
-            buyer{
-            username
-            }
-        }
-    }
-`
+async function buyBook(bookId) {
+    return await axios.post(`http://localhost:2000/book/${bookId}/buy`)
+}
 
 function Home(props) {
-    const { data, loading, error, fetchMore, refetch } = useQuery(GET_POSTS);
-    const [buyBook, { buyBookData }] = useMutation(BUY_BOOK);
     const [prevCursor, setPrevCursor] = useState(null)
+    const [data, setData] = useState([])
 
     useEffect(() => {
-        // window.addEventListener("scroll",FETCH_MORE)
-        refetch()
-
-        // return ()=> window.removeEventListener("scroll",FETCH_MORE)
-    }, [])
-    if (loading) return (<p>LOADING... </p>)
-    if (error) return (<p>{error.toString()}</p>);
-    if (!data) return (<p>Not found</p>);
-
-
-    function handleLoadMore() {
-        if (data.feed.cursorId === prevCursor) {
-            return
-        } else {
-            setPrevCursor(data.feed.cursorId)
-            console.log("load more")
-        }
-        fetchMore({
-            variables: {
-                cursorId: data.feed.cursorId,
-            },
-            updateQuery: (prev, { fetchMoreResult, ...rest }) => {
-                if (!fetchMoreResult) return prev;
-                console.log(fetchMoreResult)
-                return {
-                    ...fetchMoreResult,
-                    feed: {
-                        ...fetchMoreResult.feed,
-                        results: [
-                            ...prev.feed.results,
-                            ...fetchMoreResult.feed.results,
-                        ],
-                    },
-                };
-            },
+        axios.get('http://localhost:2000/feed', {
         })
-    }
-    if (!data.feed.isNotEmpty && !data.feed.cursorId) {
-        return <p>Noone has posted anything be the first one!</p>
-    }
-    return (
-        // <Fragment>
-        //     {data.feed.results.map(elem => <FeedBox key={elem.id}
-        //         bookId={elem.id}
-        //         volumeIdGG={elem.volumeIdGG}
-        //         sellerUsername={elem.ownedBy.username}
-        //         price={elem.price}
-        //         dateForSale={elem.dateForSale}
-        //         buyBook={buyBook}
-        //     />
-        //     )}
-        //     {data.feed.isNotEmpty && (
-        //         <button
-        //             onClick={() =>
-        //                 fetchMore({
-        //                     variables: {
-        //                         cursorId: data.feed.cursorId,
-        //                     },
-        //                     updateQuery: (prev, { fetchMoreResult, ...rest }) => {
-        //                         if (!fetchMoreResult) return prev;
-        //                         console.log(fetchMoreResult)
-        //                         return {
-        //                             ...fetchMoreResult,
-        //                             feed: {
-        //                                 ...fetchMoreResult.feed,
-        //                                 results: [
-        //                                     ...prev.feed.results,
-        //                                     ...fetchMoreResult.feed.results,
-        //                                 ],
-        //                             },
-        //                         };
-        //                     },
-        //                 })
-        //             }
-        //         >
-        //             Load More
-        //         </button>
-        //     ) || (
-        //             <p>This is the end</p>
-        //         )
-        //     }
-        // </Fragment>
-        <Fragment>
+            .then(data => data.data)
+            .then(data => setData(data))
+    }, [])
 
-        <SellPage />
-            <InfiniteScroll
+    return (
+
+        <div style={{display: "grid", gap: "1rem", boxSizing: "border-box", padding: "1rem", gridTemplateColumns: "repeat(5,1fr)"}}>
+
+            {/* <InfiniteScroll
                 loadMore={handleLoadMore}
                 hasMore={data.feed.isNotEmpty}
-                loader={<p key="loading">Loading...</p>}>
-                {data.feed.results.map(elem => <FeedBox key={elem.id}
-                    bookId={elem.id}
-                    volumeIdGG={elem.volumeIdGG}
-                    sellerUsername={elem.ownedBy.username}
-                    price={elem.price}
-                    dateForSale={elem.dateForSale}
-                    buyBook={buyBook}
-                />
-                )}
-            </InfiniteScroll>
-            {!data.feed.isNotEmpty && <p>This is the end</p>}
-        </Fragment>
+                loader={<p key="loading">Loading...</p>}> */}
+            {data.map(elem => <FeedBox key={elem.book_id}
+                bookId={elem.book_id}
+                ol_id={elem.ol_id}
+                sellerUsername={elem.owner_username}
+                price={elem.price}
+                dateForSale={elem.date_for_sale}
+                buyBook={buyBook}
+            />
+            )}
+            {/* </InfiniteScroll> */}
+            {/* {!data.feed.isNotEmpty && <p>This is the end</p>} */}
+        </div>
     );
 }
 
